refactor(app-distribution): split namespaced and modular entrypoints
